+++ tests/test_components/test_grid_spec.py
@@ -114,9 +114,11 @@ def test_autogrid_2dmaterials():
         grid_spec=td.GridSpec.auto(),
         run_time=1e-12,
     )
-    assert np.isclose(sim.volumetric_structures[0].geometry.center[2], 1, rtol=RTOL)
+    assert np.isclose(sim.volumetric_structures[0].geometry.bounding_box.center[2], 1, rtol=RTOL)
     grid_dl = sim.discretize(box.geometry).sizes.z[0]
-    assert np.isclose(sim.volumetric_structures[0].geometry.size[2], grid_dl, rtol=RTOL)
+    assert np.isclose(
+        sim.volumetric_structures[0].geometry.bounding_box.size[2], grid_dl, rtol=RTOL
+    )
 
     # now if we increase conductivity, the in-plane grid size should decrease
     sigma2 = 4.5